Promote 12 permuter score-0 hits + sort self-monitor by score
Score-0 promotions (each was a parked seed in tough_nuts/, permuter
found a byte-identical mutation; verified at SHA-1 round-trip):

- func_00104A38, func_00106D20: leaf calls
- func_001050B0
- func_0013D440, func_0013F130
- func_00164EB0, func_00164EF8, func_00164F40, func_00164F88
- func_00165EB8
- func_001A9F28, func_001A9FE0

Plus tools/self-monitor.sh now sorts the permuter-scores grid by
score ascending (lowest = closest to a match), so promotable score-0
hits surface at the top of the dashboard instead of being buried by
alphabetical func-name sort.

Park func_001F0878: permuter score-0 was a false positive; the
mutated source compiles to a smaller stack frame (-0x30 vs original
-0x40). Reseeded under the corrected permuter pipeline (see prior
commit) — re-runs may find a real hit.

Park func_0024DC40 / func_0024DC90 / func_0024DCE8: 5-arg t0-pinned
wrappers in 14DAB8 cluster. ee-gcc 2.9 emits `or rt,$0,$0` (0x25)
for `move rt, zero` but originals use `daddu rt,$0,$0` (0x2D). Same
operation, different bytes — needs a daddu-forcing idiom we don't
have yet.

+++ tough_nuts/func_001AE1B8/0AE1B8.c
@@ -1,12 +1,13 @@
+
 extern void func_00264128(int x, int y);
 extern void func_001A6E28(char *p, int y);
 extern char D_006168F8[];
-
 void func_001AE1B8(int *self, int a1, int a2)
 {
-    if (a1 != 0) {
-        func_00264128(a1, self[0] + self[0x4 / 4]);
-    }
-    self[0x4 / 4] = self[0x4 / 4] + a2;
-    func_001A6E28(D_006168F8, self[0x4 / 4]);
+  if (a1 != 0)
+  {
+    func_00264128((unsigned short) a1, self[0] + self[0x4 / 4]);
+  }
+  self[0x4 / 4] = self[0x4 / 4] + a2;
+  func_001A6E28(D_006168F8, self[0x4 / 4]);
 }
